% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{fswatch}
\alias{fswatch}
\title{Watch a File or Directory for Changes}
\format{An object of class \code{fswatch_package} of length 1.}
\usage{
fswatch(path, callback, check_interval = 500)

## id <- fswatch(path, callback, check_interval = 500)
## fswatch$cancel(id)
## fswatch$list()
}
\arguments{
\item{path}{Path to watch. It can be a directory or a file.
Directories are watched recursively.}

\item{callback}{Callback function to call when \code{path} is
updated, i.e. the modification time of the file, or a file in the
directory changed. The function will be called with one argument:
a character vector of the changed files and directories.}

\item{check_interval}{How long to wait between checks for changes,
in milliseconds. Defaults to half a second.}
}
\value{
The ID of the watcher. This can be used to cancel it later.
}
\description{
Get notification is a file or any file in a directory changes. It uses
the fswatch tool (see \url{https://github.com/emcrisostomo/fswatch})
and works on major platforms.
}
\section{Additional methods}{


\code{fswatch$list()} lists all watchers.

\code{fswatch$cancel(id} cancels a watcher.
}
\examples{
id <- fswatch(tempdir(), callback = function(f) print(f))
tmp <- tempfile()
tmp
cat("hello", file = tmp)
Sys.sleep(1)
file.remove(tmp)
Sys.sleep(1)

fswatch$list()
fswatch$cancel(id)
fswatch$list()
}
\keyword{datasets}

