% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ fswatch }
\alias{ fswatch }
\alias{ fswatch }
\title{ Watch a File or Directory for Changes }
\description{
Get notification is a file or any file in a directory changes. It uses the
  fswatch tool (see &lt;https://github.com/emcrisostomo/fswatch&gt;) and works on
  major platforms.
}
